# public/swagger.yaml
openapi: 3.0.3
info:
  title: Vinci Barbearia API
  description: |
    API completa para gerenciamento de barbearia incluindo agendamentos, barbeiros, clientes, produtos, planos e webhooks.

    ## Autenticação
    A maioria dos endpoints requer autenticação via Bearer Token no header:
    ```
    Authorization: Bearer SEU_TOKEN_AQUI
    ```

    ## Formato de Datas
    - Datas são no formato brasileiro: DD/MM/YYYY ou YYYY-MM-DD
    - Horários são no formato: HH:MM (24h)

    ## Webhooks
    O sistema suporta dois tipos de webhooks:
    1. **Webhook Global**: Configurado na tabela `configuracoes`, dispara para todos os eventos
    2. **Webhooks de Barbeiros**: Webhooks personalizados por profissional

  version: 3.0.0
  contact:
    name: Vinci Barbearia
    url: https://vincebarbearia.vercel.app
servers:
  - url: https://vincebarbearia.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticação da API
...
paths:
...
  /barbeiros/listar:
    get:
      tags:
        - Barbeiros
      summary: Listar todos os barbeiros
      description: |
        Retorna lista completa de barbeiros com estatísticas.

        **Retorna:**
        - Lista de barbeiros ordenados por quantidade de atendimentos
        - Próximo barbeiro do rodízio
        - Estatísticas de cada barbeiro
      parameters:
        - name: ativo
          in: query
          schema:
            type: boolean
            default: true
          description: Filtrar apenas barbeiros ativos
      responses:
        '200':
          description: Lista de barbeiros retornada
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  proximo_rodizio:
                    type: object
                  barbeiros:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        nome:
                          type: string
                        telefone:
                          type: string
                        email:
                          type: string
                        especialidades:
                          type: string
                        ativo:
                          type: boolean
                        estatisticas:
                          type: object
                          properties:
                            total_atendimentos:
                              type: integer
                            atendimentos_hoje:
                              type: integer
                            total_concluidos:
                              type: integer
...
  /barbeiros/meu-faturamento:
    get:
      tags:
        - Barbeiros
      summary: Meu faturamento do barbeiro
      description: Retorna faturamento do barbeiro com opção de período
      security:
        - BearerAuth: []
      parameters:
        - name: barbeiro_nome
          in: query
          required: true
          schema:
            type: string
          description: Nome do barbeiro
          example: "Hiago"
        - name: periodo
          in: query
          schema:
            type: string
            default: hoje
            enum: [hoje, semana, mes, ano]
          description: Período do faturamento
      responses:
        '200':
          description: Faturamento retornado
...